pyproject.toml
[project]
name = "mixnet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=45.0.4",
    "grpcio>=1.73.0",
    "grpcio-tools>=1.73.0",
    "pydantic>=2.11.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
]
